[#157391555]: Fix MR margin calculation with H&S
- Don't include H&S in actual margin dimension (don't affect
  calculated margin results if there are H&S)
- Support with tests
