+++ tests/integration/fixtures/__init__.py
@@ -121,3 +121,4 @@ BBC_NEWS = _load('bbc-news.json')
 AGE_X_ACCRPIPE = _load('age-x-accrpipe.json')
 CAT_X_MR_X_CAT_MISSING = _load('cat-x-mr-x-cat-missing.json')
 MR_X_CA_HS = _load('mr-x-ca-with-hs.json')
+CA_X_MR_WEIGHTED_HS = _load('ca-x-mr-weighted-hs.json')
